Changed type to long

+++ src/ZendeskApi.Client/Models/AuditEvents/TicketSharingEvent.cs
@@ -5,7 +5,7 @@ namespace ZendeskApi.Client.Models.AuditEvents
     public class TicketSharingEvent : AuditEvent
     {
         [JsonProperty("agreement_id")]
-        public int AgreementId { get; set; }
+        public long AgreementId { get; set; }
         [JsonProperty("action")]
         public string Action { get; set; }
     }
